+++ packages/contact-center/cc-widgets/src/index.ts
@@ -1,5 +1,6 @@
 import {StationLogin} from '@webex/cc-station-login';
 import {UserState} from '@webex/cc-user-state';
+import {IncomingTask, TaskList} from '@webex/cc-task';
 import store from '@webex/cc-store';
 
-export {StationLogin, UserState, store};
+export {StationLogin, UserState, IncomingTask, TaskList, store};