+++ packages/contact-center/station-login/src/station-login/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import store from '@webex/cc-store';
-import {observer} from 'mobx-react';
+import {observer} from 'mobx-react-lite';
 
 import StationLoginPresentational from './station-login.presentational';
 import {useStationLogin} from '../helper';