+++ packages/contact-center/store/src/store.ts
@@ -12,15 +12,31 @@ import {
 } from './store.types';
 
 class Store implements IStore {
+  private static instance: Store;
   teams: Team[] = [];
   loginOptions: string[] = [];
   cc: IContactCenter;
   logger: ILogger;
   idleCodes: IdleCode[] = [];
   agentId: string = '';
+  selectedLoginOption: string = '';
 
   constructor() {
     makeAutoObservable(this, {cc: observable.ref});
+  }
+
+  public static getInstance(): Store {
+    if (!Store.instance) {
+      console.log('Creating new store instance');
+      Store.instance = new Store();
+    }
+
+    console.log('Returning store instance');
+    return Store.instance;
+  }
+
+  setSelectedLoginOption(option: string): void {
+    this.selectedLoginOption = option;
   }
 
   registerCC(webex: WithWebex['webex']): Promise<void> {
@@ -41,13 +57,12 @@ class Store implements IStore {
   }
 
   init(options: InitParams): Promise<void> {
-    if('webex' in options) {
+    if ('webex' in options) {
       // If devs decide to go with webex, they will have to listen to the ready event before calling init
-      // This has to be documented 
+      // This has to be documented
       return this.registerCC(options.webex);
     }
     return new Promise((resolve, reject) => {
-
       const timer = setTimeout(() => {
         reject(new Error('Webex SDK failed to initialize'));
       }, 6000);
@@ -55,22 +70,23 @@ class Store implements IStore {
       const webex = Webex.init({
         config: options.webexConfig,
         credentials: {
-          access_token: options.access_token
-        }
+          access_token: options.access_token,
+        },
       });
-  
+
       webex.once('ready', () => {
         clearTimeout(timer);
-        this.registerCC(webex).then(() => {
-          resolve();
-        })
-        .catch((error) => {
-          reject(error);
-        })
-      })
+        this.registerCC(webex)
+          .then(() => {
+            resolve();
+          })
+          .catch((error) => {
+            reject(error);
+          });
+      });
     });
   }
 }
 
-const store = new Store();
+const store = Store.getInstance();
 export default store;